Fix image uploads by allowing non-String bodies in StripeRequest
The issue boiled down to StripeRequest, where the `body` parameter was always expected to be a human-friendly String.
Real image data is typically not representable as UTF-8 strings, so the forced conversion would always fail this way or another.

The change from String to LosslessHTTPBodyRepresentable does not affect library code and should not affect users either –
any String that was used before automatically complies to this more generic protocol.